--- src/main/java/com/likelion/songyeechallenge/controller/AdminController.java
@@ -1,18 +1,0 @@
-package com.likelion.songyeechallenge.controller;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-@RequestMapping("admin")
-public class AdminController {
-    private static final Logger logger = LoggerFactory.getLogger(AdminController.class);
-
-    @GetMapping
-    public String adminHome(){
-        return "admin";
-    }
-}
