+++ src/main/java/com/likelion/songyeechallenge/controller/MyPageController.java
@@ -1,22 +1,14 @@
 package com.likelion.songyeechallenge.controller;
 
-import com.likelion.songyeechallenge.config.JwtTokenProvider;
-import com.likelion.songyeechallenge.domain.challenge.Challenge;
-import com.likelion.songyeechallenge.domain.user.User;
-import com.likelion.songyeechallenge.domain.user.UserRepository;
 import com.likelion.songyeechallenge.service.MyPageService;
 import com.likelion.songyeechallenge.service.ReviewService;
 import com.likelion.songyeechallenge.web.dto.*;
-import com.likelion.songyeechallenge.service.UserService;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/v1/mypage")
@@ -56,31 +48,6 @@ public class MyPageController {
         return myPageService.findMyReview(jwtToken);
     }
 
-    @GetMapping("/mission")
-    public List<MyMissionResponseDto> getUserMissions(@RequestHeader("Authorization") String authorizationHeader) {
-        String jwtToken = authorizationHeader.replace("Bearer ", "");
-        return myPageService.findMyChallengeAndMission(jwtToken);
-    }
-
-    @GetMapping("/info")
-    public List<UserInfoDto> getUserInfo(@RequestHeader("Authorization") String authorizationHeader) {
-        String jwtToken = authorizationHeader.replace("Bearer ", "");
-        return myPageService.showMyInfo(jwtToken);
-    }
-
-
-    @DeleteMapping("/delete-user")
-    public ResponseEntity<String> deleteUser(@RequestHeader("Authorization") String authorizationHeader){
-        String jwtToken = authorizationHeader.replace("Bearer ", "");
-        try {
-            myPageService.deleteAccount(jwtToken);
-            return ResponseEntity.ok("User account deleted successfully.");
-        } catch (Exception e) {
-            // 실패한 경우 예외를 캐치하여 실패 응답을 반환
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to delete user account.");
-        }
-    }
-
     @DeleteMapping("/review/delete/{id}")
     public ResponseEntity<String> deleteReview(@PathVariable Long id, @RequestHeader("Authorization") String authorizationHeader) {
         String jwtToken = authorizationHeader.replace("Bearer ", "");
@@ -108,9 +75,34 @@ public class MyPageController {
         }
     }
 
+    @GetMapping("/mission")
+    public List<MyMissionResponseDto> getUserMissions(@RequestHeader("Authorization") String authorizationHeader) {
+        String jwtToken = authorizationHeader.replace("Bearer ", "");
+        return myPageService.findMyChallengeAndMission(jwtToken);
+    }
+
     @PostMapping("/mission/{missionId}/{challengeId}")
     public boolean isCompleteMission(@PathVariable("missionId") Long missionId, @RequestHeader("Authorization") String authorizationHeader) {
         String jwtToken = authorizationHeader.replace("Bearer ", "");
         return myPageService.isCompleteMission(missionId, jwtToken);
+    }
+
+    @GetMapping("/info")
+    public List<UserInfoDto> getUserInfo(@RequestHeader("Authorization") String authorizationHeader) {
+        String jwtToken = authorizationHeader.replace("Bearer ", "");
+        return myPageService.showMyInfo(jwtToken);
+    }
+
+
+    @DeleteMapping("/delete-user")
+    public ResponseEntity<String> deleteUser(@RequestHeader("Authorization") String authorizationHeader){
+        String jwtToken = authorizationHeader.replace("Bearer ", "");
+        try {
+            myPageService.deleteAccount(jwtToken);
+            return ResponseEntity.ok("User account deleted successfully.");
+        } catch (Exception e) {
+            // 실패한 경우 예외를 캐치하여 실패 응답을 반환
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to delete user account.");
+        }
     }
 }
