+++ src/main/java/com/likelion/songyeechallenge/domain/like/LikeRepository.java
@@ -1,4 +1,4 @@
-package com.likelion.songyeechallenge.domain.likes;
+package com.likelion.songyeechallenge.domain.like;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
