+++ src/main/java/com/likelion/songyeechallenge/service/UserService.java
@@ -5,14 +5,6 @@ import com.likelion.songyeechallenge.domain.auth.AuthRepository;
 import com.likelion.songyeechallenge.config.CustomUserDetails;
 import com.likelion.songyeechallenge.config.JwtTokenProvider;
 import com.likelion.songyeechallenge.config.dto.AuthResponseDto;
-import com.likelion.songyeechallenge.domain.challenge.Challenge;
-import com.likelion.songyeechallenge.domain.challenge.ChallengeRepository;
-import com.likelion.songyeechallenge.domain.likes.Like;
-import com.likelion.songyeechallenge.domain.likes.LikeRepository;
-import com.likelion.songyeechallenge.domain.mission.Mission;
-import com.likelion.songyeechallenge.domain.mission.MissionRepository;
-import com.likelion.songyeechallenge.domain.review.Review;
-import com.likelion.songyeechallenge.domain.review.ReviewRepository;
 import com.likelion.songyeechallenge.domain.user.Role;
 import com.likelion.songyeechallenge.domain.user.User;
 import com.likelion.songyeechallenge.domain.user.UserRepository;
@@ -26,9 +18,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.Set;
-
 @Slf4j
 @RequiredArgsConstructor
 @Service
@@ -36,10 +25,6 @@ public class UserService {
 
     private final UserRepository userRepository;
     private final AuthRepository authRepository;
-    private final ReviewRepository reviewRepository;
-    private final LikeRepository likeRepository;
-    private final ChallengeRepository challengeRepository;
-    private final MissionRepository missionRepository;
     private final PasswordEncoder passwordEncoder;
     private final JwtTokenProvider jwtTokenProvider;
 
@@ -104,46 +89,5 @@ public class UserService {
             return newAccessToken;
         }
         return null;
-    }
-
-    @Transactional(readOnly = true)
-    public User getUserProfile(Long userId) {
-        return userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 사용자를 찾을 수 없습니다. ID: " + userId));
-    }
-
-    @Transactional
-    public void deleteUser(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 사용자를 찾을 수 없습니다. ID: " + userId));
-
-        // 사용자와 연관된 리뷰 삭제
-        List<Review> userReviews = reviewRepository.findByUser(userId);
-        for (Review review : userReviews) {
-            // 리뷰와 관련된 다양한 데이터 삭제 (예시: 좋아요 등)
-            List<Like> reviewLikes = (List<Like>) likeRepository.findByReview(userId);
-            likeRepository.deleteAll(reviewLikes);
-
-            // 리뷰 삭제
-            reviewRepository.delete(review);
-        }
-
-        // 사용자와 연관된 좋아요 데이터 삭제
-        List<Like> userLikes = (List<Like>) likeRepository.findByUser(userId);
-        likeRepository.deleteAll(userLikes);
-
-        // 사용자와 연관된 챌린지 데이터 삭제
-        Set<Challenge> userChallenges = user.getChallenges();
-        for (Challenge challenge : userChallenges) {
-            // 챌린지와 관련된 다양한 데이터 삭제
-            challengeRepository.delete(challenge);
-        }
-
-        // 사용자와 연관된 미션 데이터 삭제
-        List<Mission> userMissions = missionRepository.findByChallengeId(userId);
-        missionRepository.deleteAll(userMissions);
-
-        // 사용자 삭제
-        userRepository.delete(user);
     }
 }
