--- src/main/resources/templates/admin.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html
-    lang = "en"
-    xmlns:th="http://www.thymeleaf.org"
-    xmlns:sec="http://www.w3.org/1999/xhtml"
->
-
-<head>
-    <meta charset = "UTF-8" />
-    <title>Simple Admin Home</title>
-</head>
-<body>
-Hello Admin!
-<form th:action="@{/user/logout}" method="post">
-    <input type="submit" th:value="로그아웃" />
-</form>
-</body>
-</html>
