--- src/main/resources/templates/home.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.ofg">
-<head>
-    <meta charset="UTF-8">
-    <title>Home</title>
-</head>
-<body>
-<div class="container">
-    <p>회원 기능</p>
-    <div>
-        <a href="/user/signup">회원가입</a>
-        <a href="/user/login">로그인</a>
-    </div>
-</div>
-</body>
-</html>
