--- src/main/resources/templates/signup.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>회원가입</title>
-</head>
-<body>
-<div class="container">
-        <!-- 에러 메시지 출력 -->
-    <div th:if="${param.error}">
-        <p style="color: red;">회원가입에 실패했습니다. 다시 시도해주세요.</p>
-    </div>
-    <form method="post" action="/user/signup">
-        <div class="form_group">
-            <label for="name">이름</label>
-            <input type="text" id="name" name="name" placeholder="이름을 입력하세요.">
-            <label for="name">아이디</label>
-            <input type="text" id="user_id" name="user_id" placeholder="아이디를 입력하세요.">
-            <label for="email">이메일</label>
-            <input type="email" id="email" name="email" placeholder="이메일을 입력하세요.">
-            <label for="password">비밀번호</label>
-            <input type="password" id="password" name="password" placeholder="비밀번호를 입력하세요.">
-            <label for="major">전공</label>
-            <input type="text" id="major" name="major" placeholder="전공을 입력하세요.">
-            <label for="student_id">학번</label>
-            <input type="number" id="student_id" name="student_id" placeholder="학번을 입력하세요.">
-            <button type="submit">회원가입</button>
-        </div>
-    </form>
-</div>
-</body>
-</html>
